Decode unicode types in decode_binary
The test in decode_binary for six.text_type was incorrect as that includes
unicode type in Python 2 which should actually be decoded.

When the type is string_types we now properly check only for basestring and
str in Python 2 and Python 3 respectively and return the given blob without
making an attempt to decode.

(cherry picked from commit 0439d8a17d181a2546f2f7cb2d71a04bbb13b186)